Deployment test: -> Added scripts to help the docker-compose run, the docker-compose now runs the scripts and then runs a migrations docker image that will update the database to match any chnages since the scripts were created. -> Updated the docker-compose with an initialization folder that runs scripts the first time the postgres db is created.
Project-Tech-Shop:
-> Added Launch settings for a tesdting startup option
-> Fixed issue with the Crazy Page where the simpleGame was broken from the AJAX request for basket at the top of the image.
-> chnaged logs created from running to be named Dev if run from Dev environment
-> Added appsettings for Test environment
-> Updated dockerfile to remove not needed service
-> Updated Startup for specific Test environment config

Project.Tech.Shop.Services.Contact:
-> Added Contact Service for customer service tools
Note: Service is still empty for the time being.

Project.Tech.Shop.Services.Migrations.Startup:
-> Added dockerfile
-> Added Appsettings for the dev and normal environment
-> Added Environemtn variable to the launchsettings.json to set environment to development, other it is production
-> changed program.cs to allow automatic updating of all the databases rather then having the manually do it.

Project.Tech.Shop.Web.Test.Playwright:
-> Added playwright Tests with GlobalSetup class to sort out the work to set up the running of the project to start testing playwright.
-> Added pages for Login and registration Page
-> Added Login, mainPage and Registration Tests
-> All test work off the PlaywrightTestBase, this adds the ability to keep naming conventions for tests, create screnshots, add tags, and finally create a report containing all the information.
-> Added ReportManager.cs Class to handle the report creation

Project.Tech.Shop.Services.Products:
-> Made general changes to the Products db entities to work better with the app, this includes adding basket details.

Project.Tech.Shop.Services.Transactions:
-> Deleted Transactions as it has been integrated into Products Service

Project.Tech.Shop.Services.Common in Test:
-> Updated to match needs for some tests and product changes

Finally:
Added ignore to the gitignore for the reports that created by Playwright

+++ application_code/test/Project.Tech.Shop.Services.Common/TestProductsContextDatabaseFactory.cs
@@ -23,12 +23,12 @@ namespace Project.Tech.Shop.Tests.Common
                 Name = fixture.Create<string>(),
                 Description = fixture.Create<string>(),
                 Price = fixture.Create<decimal>(),
-                Category = fixture.Create<Category>(),
+                Category = (int)fixture.Create<Category>(),
                 StockLevel = fixture.Create<int>(),
                 Brand = fixture.Create<string>()
             };
 
-            if (product.Category == Category.Laptop)
+            if (product.Category == (int)Category.Laptop)
             {
                 product.Series = fixture.Create<string>();
                 product.ProcessorType = fixture.Create<string>();
@@ -87,7 +87,7 @@ namespace Project.Tech.Shop.Tests.Common
                 Name = fixture.Create<string>(),
                 Description = fixture.Create<string>(),
                 Price = fixture.Create<decimal>(),
-                Category = fixture.Create<Category>(),
+                Category = (int)fixture.Create<Category>(),
                 StockLevel = fixture.Create<int>(),
                 Brand = fixture.Create<string>()
             };
